完善游戏

+++ src/main/java/com/rhino/bjl/server/IMainMessage.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface IMainMessage {
     HashMap<String, Object> findParamMsgByUserId(String userId);
 
-    boolean updateUserMoneyByUserId(String userId, int money);
+    boolean updateUserMoneyByUserId(String userId, float money);
 
     boolean updateZhuangMoneyByUserId(String userId);
